Semi-working ramp up and down.
State and level are not in 'noinit' variables so the short-clicks can be
handled without reloading the state to eeprom. This means the ramps can be done
without constantly writing to eeprom which will make things simpler and safer
(less wear).

The downside is that the sram retention delay might be a bit unreliable. But it
seems to be more than long enough for now. Also, it won't be possible to long
click from an up ramp. It will start ramping again from the previous start
position. This doesn't seem too bad.

This commit also includes eeprom read/write code which is untested.