Fixed nasty bug with cell level reading on click.
The MCU keeps running long enough on the capacitor to read that there is no
voltage from the cell anymore. This was messing up code which is based on that
reading:
- LVP would drop the output by a lot, visible as lower output after a click
  with another change to use that value.
- Because the gate charge is voltage compensated, a very low Vcc could give a
  very large charge and be the cause of annoying bright flashes. Need more
  testing to confirm this but the flashes to appear to be gone.

The fix, for now, is to allow the reading to decrease by one unit at most for
every WDT interrupt. The MCU won't run long enough for it to drop much and this
won't affect normal use.