Fix cell level reading on startup.
The fix of c088f7da59 was causing a long delay before getting the right level
when starting up with a near empty cell. This, in turn, caused the light to
slowly ramp back up to the correct output after a click. This commit fixes that
problem by not putting a decrease limit on the first readout.

The weird initial value is to avoid overflowing a signed 8-bit int when
ADC_CELL_LOWEST is < 128. Otherwise, LVP blows up and the light never comes on.